+++ crates/prelude/src/scheduling.rs
@@ -46,3 +46,8 @@ pub fn wait_indefinitely() -> ! {
         wait_for_callback();
     }
 }
+
+/// Executes a breakpoint.
+pub fn breakpoint() {
+    unsafe { api::breakpoint() };
+}