+++ crates/api-desc/CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+### Minor
+
+- Add scheduling::breakpoint
+
 ### Patch
 
 - Test that link names are unique
@@ -13,3 +17,5 @@
 - Add copyright header on generated files
 
 ## 0.1.0
+
+<!-- Update PR number to skip CHANGELOG.md test: #0 -->