+++ crates/board/CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+### Minor
+
+- Add breakpoint in Api
+
 ## 0.1.2
 
 ### Patch
@@ -13,3 +17,5 @@
 - Build `docs.rs` with `std` feature
 
 ## 0.1.0
+
+<!-- Update PR number to skip CHANGELOG.md test: #0 -->