+++ crates/interpreter/CHANGELOG.md
@@ -15,3 +15,5 @@
 - Panic if `start` calls a host function
 
 ## 0.1.0
+
+<!-- Update PR number to skip CHANGELOG.md test: #0 -->