+++ crates/scheduler/CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+### Minor
+
+- Support scheduling::breakpoint
+
 ## 0.1.1
 
 ### Patch
@@ -7,3 +11,5 @@
 - Use `logger::panic!` instead of `core::panic!`
 
 ## 0.1.0
+
+<!-- Update PR number to skip CHANGELOG.md test: #0 -->